+++ app/javascript/src/views/ArticleModal/index.js
@@ -1,7 +1,7 @@
 import { motion } from "framer-motion";
 import React, { useCallback, useEffect, useLayoutEffect, useRef } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
-import CaseStudyArticleNew from "../CaseStudyArticleNew";
+import CaseStudyArticle from "../CaseStudyArticle";
 
 const backdropClassName = `
   fixed
@@ -77,7 +77,7 @@ export default function ArticleModal() {
         data-testid="articleModal"
       >
         <div className="overflow-y-scroll h-full" id="article-scrollable-area">
-          <CaseStudyArticleNew />
+          <CaseStudyArticle />
         </div>
       </motion.div>
     </motion.div>
