+++ app/javascript/src/views/CaseStudyArticle/components/ArticleContent.js
