+++ app/javascript/src/views/CaseStudyArticle/components/Heading.js
