+++ app/javascript/src/views/CaseStudyArticle/components/KeyTakeaways.js
