+++ app/javascript/src/views/CaseStudyArticle/components/LimitedContentSignup.js
