+++ app/javascript/src/views/CaseStudyArticle/components/Paragraph.js
