+++ app/javascript/src/views/CaseStudyArticle/components/ScrollIndicator.js
