+++ app/javascript/src/views/CaseStudyArticle/components/ShareArticleButton.js
