+++ app/javascript/src/views/CaseStudyArticle/index.js
