+++ app/javascript/src/views/CaseStudyArticle/queries/article.gql
