+++ app/javascript/src/views/CaseStudyArticle/queries/heroFields.gql
