+++ app/javascript/src/views/FreelancerProfile/queries/fragments/specialistFields.gql
@@ -1,4 +1,4 @@
-#import 'src/views/CaseStudyArticleNew/queries/heroFields.gql'
+#import 'src/views/CaseStudyArticle/queries/heroFields.gql'
 
 fragment SpecialistFields on Specialist {
   id
